requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
anyio==3.7.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==23.1.0
azure-core==1.28.0
azure-storage-blob==12.17.0
backoff==2.2.1
blinker==1.6.2
boilerpy3==1.0.6
cachetools==5.3.1
canals==0.2.2
cattrs==23.1.2
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.2.0
chromadb==0.3.29
click==8.1.4
clickhouse-connect==0.6.6
coloredlogs==15.0.1
constants==0.6.0
cryptography==41.0.2
dataclasses-json==0.5.9
decorator==5.1.1
docopt==0.6.2
duckdb==0.8.1
et-xmlfile==1.1.0
Events==0.4
exceptiongroup==1.1.2
farm-haystack==1.18.1
fastapi==0.85.1
filelock==3.12.2
filetype==1.2.0
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.32
greenlet==2.0.2
h11==0.14.0
hnswlib==0.7.0
httptools==0.6.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
inflect==7.0.0
isodate==0.6.1
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.18.2
jsonschema-specifications==2023.6.1
langchain==0.0.231
langchainplus-sdk==0.0.20
lazy-imports==0.3.1
lxml==4.9.3
lz4==4.3.2
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
mdurl==0.1.2
monotonic==1.6
more-itertools==9.1.0
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
num2words==0.5.12
numexpr==2.8.4
numpy==1.25.1
olefile==0.46
onnxruntime==1.15.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
overrides==7.3.1
packaging==23.1
pandas==2.0.3
pdf2image==1.16.3
pdfminer.six==20221105
Pillow==9.5.0
platformdirs==3.8.1
posthog==3.0.1
prompthub-py==4.0.0
protobuf==4.23.4
pulsar-client==3.2.0
pyarrow==12.0.1
pycparser==2.21
pydantic==1.10.11
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pypandoc==1.11
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.27
python-pptx==0.6.21
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
quantulum3==0.9.0
rank-bm25==0.2.2
referencing==0.29.1
regex==2023.6.3
requests==2.31.0
requests-cache==0.9.8
rich==13.4.2
rpds-py==0.8.10
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==2.0.18
sseclient-py==1.7.2
starlette==0.20.4
streamlit==1.24.1
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
threadpoolctl==3.2.0
tiktoken==0.4.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
transformers==4.30.1
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
unstructured==0.8.1
url-normalize==1.4.3
urllib3==2.0.3
uvicorn==0.22.0
uvloop==0.17.0
validators==0.20.0
watchdog==3.0.0
watchfiles==0.19.0
websockets==11.0.3
xlrd==2.0.1
XlsxWriter==3.1.2
yarl==1.9.2
zipp==3.16.0
zstandard==0.21.0
